Fix routes

+++ src/controllers/costController.js
@@ -4,7 +4,6 @@ const Cost = require('../models/cost');
 const Debt = require('../models/debt');
 
 async function index(request, response) {
-
     const { trip, user } = request.params;
 
     try {
@@ -14,14 +13,11 @@ async function index(request, response) {
 
         return response.status(200).json(cost);
     } catch (error) {
-        return response.status(500).json(
-            message = error,
-        );
+        return response.status(500).json({ message: error });
     }
 };
 
 async function create(request, response) {
-
     const id = uuidv4();
     const {
         description,
@@ -59,14 +55,11 @@ async function create(request, response) {
 
         return response.status(200).json(cost);
     } catch (error) {
-        return response.status(500).json(
-            message = error,
-        );
+        return response.status(500).json({ message: error });
     }
 };
 
 async function read(request, response) {
-
     const { id } = request.params;
 
     try {
@@ -81,50 +74,83 @@ async function read(request, response) {
         const debts = await Debt.findAll({
             where: { cost: cost.id },
         });
-        return response.status(200).json({ cost: cost, debts: debts });
 
+        return response.status(200).json({ cost: cost, debts: debts });
     } catch (error) {
-        return response.status(500).json(
-            message = error,
-        );
+        return response.status(500).json({ message: error });
     }
 };
 
 async function edit(request, response) {
-
     const { id } = request.params;
-    const { description, value, dtcost } = request.body;
+    const {
+        description, value, category, subcategory, dtcost, participants,
+    } = request.body;
 
     try {
+        const debts = await Debt.findAll({
+            where: { cost: id, settled: true }
+        });
+
+        if (debts !== null) {
+            return response.status(406).json({
+                debts,
+                message: 'Custo possui dívidas já pagas!',
+            })
+        }
+
+        await Debt.destroy({ where: { cost: id }});
         const cost = await Cost.update({
             description,
             value,
+            category,
+            subcategory,
             dtcost,
         }, {
             where: { id },
         });
+        
         if (cost[0] === 0) {
             return response.status(404).json({
                 message: 'Custo não encontrado!',
             });
         }
+
+        if (participants.length > 0) {
+            participants.forEach(async (participant) => {
+                await Debt.create({
+                    user: participant,
+                    cost: cost.id,
+                    value: cost.value / (participants.length + 1),
+                    settled: false,
+                });
+            });
+        }
+
         return response.status(200).json({
             data: cost[0],
             message: 'Custo atualizado com sucesso!',
         });
-
     } catch (error) {
-        return response.status(500).json(
-            message = error,
-        )
+        return response.status(500).json({ message: error });
     }
 };
 
 async function remove(request, response) {
-
     const { id } = request.params;
 
     try {
+        const debts = await Debt.findAll({
+            where: { cost: id, settled: true }
+        });
+
+        if (debts !== null) {
+            return response.status(406).json({
+                debts,
+                message: 'Custo possui dívidas já pagas!',
+            })
+        }
+
         await Debt.destroy({ where: { cost: id }});
         const cost = await Cost.destroy({ where: { id } });
 
@@ -133,15 +159,13 @@ async function remove(request, response) {
                 message: 'Custo não encontrado!',
             });
         }
+        
         return response.status(200).json({
             data: cost[0],
             message: 'Custo apagado com sucesso!',
         });
-
     } catch (error) {
-        return response.status(500).json(
-            message = error,
-        )
+        return response.status(500).json({ message: error });
     }
 };
